Resolved the comments
